+++ Tp2_Grupo30/src/main/java/ar/edu/unju/fi/service/IConsejoService.java
@@ -1,19 +1,21 @@
 package ar.edu.unju.fi.service;
 
+import java.io.IOException;
+import java.util.List;
+
 import org.springframework.web.multipart.MultipartFile;
 
-import ar.edu.unju.fi.listas.ListaConsejo;
 import ar.edu.unju.fi.model.Consejo;
 
 public interface IConsejoService {
 
-	public ListaConsejo getConsejos();
+	public List<Consejo> getConsejos();
 	
-	public void addConsejo(Consejo consejo,MultipartFile imagen);
+	public void addConsejo(Consejo consejo,MultipartFile imagen) throws IOException;
 	
 	public Consejo getConsejo(int id);
 	
-	public void updateConsejo(Consejo consejoModificado,MultipartFile imagen);
+	public void updateConsejo(Consejo consejoModificado,MultipartFile imagen) throws IOException;
 	
 	public void deleteConsejo(int id);
